+++ src/commands/index.ts
@@ -1,6 +1,5 @@
 import { ExtensionContext } from "vscode";
 import { registerAuthCommands } from "./auth";
-import { registerCodePenCommands } from "./codepen";
 import { registerCommentCommands } from "./comments";
 import { registerDirectoryCommands } from "./directory";
 import { registerEditorCommands } from "./editor";
@@ -8,7 +7,6 @@ import { registerFileCommands } from "./file";
 import { registerFollowCommands } from "./follow";
 import { registerGistCommands } from "./gist";
 import { registerGistLogCommands } from "./gistLog";
-import { registerPlaygroundCommands } from "./playground";
 import { registerScratchCommands } from "./scratch";
 import { registerTourCommands } from "./tour";
 
@@ -20,8 +18,6 @@ export function registerCommands(context: ExtensionContext) {
   registerGistCommands(context);
   registerGistLogCommands(context);
   registerFileCommands(context);
-  registerPlaygroundCommands(context);
-  registerCodePenCommands(context);
   registerDirectoryCommands(context);
   registerTourCommands(context);
   registerScratchCommands(context);