+++ src/constants.ts
@@ -4,13 +4,12 @@ export const EXTENSION_ID = "vsls-contrib.gistfs";
 export const FS_SCHEME = "gist";
 export const INPUT_SCHEME = "input";
 
-export const PLAYGROUND_FILE = "playground.json";
+export const SWING_FILE = "codeswing.json";
 export const UNTITLED_SCHEME = "untitled";
 
 export const URI_PATTERN = /https?:\/\/(www\.)?[-a-zA-Z0-9@:%._\+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_\+.~#?&//=]*)/;
 
 export const ZERO_WIDTH_SPACE = "‎‎\u200B";
-export const TEMP_GIST_ID = "temp";
 
 export const DIRECTORY_SEPARATOR = "/";
 export const ENCODED_DIRECTORY_SEPARATOR = "---";